+++ Changelog.md
@@ -1,4 +1,9 @@
+## [1.0.2] - 2019-11-04
+### Changed
+- Changed the mail address of the author.
+- Fixed a compatibility bug with Python 2.
+
 ## [1.0.1] - 2019-09-10
 ### Added
-- Corrected version numbering
-- Added some author information
+- Corrected version numbering.
+- Added some author information.